fix tests: - Migrate dynamic tests to parameterized tests - Change order of delete operations in databaseUtilService
