+++ src/test/java/de/tum/in/www1/artemis/util/DatabaseUtilService.java
@@ -121,13 +121,10 @@ public class DatabaseUtilService {
         modelingSubmissionRepo.deleteAll();
         textSubmissionRepo.deleteAll();
         programmingSubmissionRepo.deleteAll();
-        exerciseRepo.deleteAll();
-        participationRepo.deleteAll();
-        programmingExerciseStudentParticipationRepo.deleteAll();
-        templateProgrammingExerciseParticipationRepo.deleteAll();
-        solutionProgrammingExerciseParticipationRepo.deleteAll();
         submissionRepository.deleteAll();
+        programmingExerciseRepository.deleteAll();
         participationRepo.deleteAll();
+        exerciseRepo.deleteAll();
         courseRepo.deleteAll();
         userRepo.deleteAll();
         assertThat(resultRepo.findAll()).as("result data has been cleared").isEmpty();
@@ -135,7 +132,6 @@ public class DatabaseUtilService {
         assertThat(exerciseRepo.findAll()).as("exercise data has been cleared").isEmpty();
         assertThat(userRepo.findAll()).as("user data has been cleared").isEmpty();
         assertThat(participationRepo.findAll()).as("participation data has been cleared").isEmpty();
-        assertThat(programmingExerciseRepository.findAll()).as("programming exercise data has been cleared").isEmpty();
         assertThat(testCaseRepository.findAll()).as("test case data has been cleared").isEmpty();
     }
